refactor(app): ヘッダーに関連するコンポーネントを1つのファイルに集約
ヘッダーを修正する場合、どのファイルを修正するか分かりやすくするため。

re #44
