--- app/app/(reception)/components/Header/nav-links.ts
@@ -1,18 +1,0 @@
-export const navLinks = [
-  {
-    label: "ユーザー検索",
-    href: "/user-search",
-  },
-  {
-    label: "利用履歴",
-    href: "/usage-logs",
-  },
-  {
-    label: "レポート",
-    href: "/reports",
-  },
-  {
-    label: "カード (NFC) 登録",
-    href: "/nfc-registration",
-  },
-];
